+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add --json flag.
 - Add --user option.
 - Add --batch-size option. 
+- Add --include-drafts flag.
 - Fix calculation of number of review and exclude author.
 - Fix calculation of number of approvals